change computation for branch 1

+++ Samples/Desktop/D3D12Raytracing/src/D3D12RaytracingMultipleDiffTexturedObjects_SER/RayTracingHlslCompat.h
@@ -33,7 +33,7 @@ struct SceneConstantBuffer
 struct CubeConstantBuffer
 {
     XMFLOAT4 albedo;
-    float materialID;
+    uint32_t materialID;
 };
 
 struct Vertex
